refactor(styles): make default-styles.json the single source of defaults
Per review: the hand-written DEFAULT_STYLES literal is gone; the default
preset json moves into src and is parsed strictly at boot. The preset picker
serves "default" from the bundle instead of fetching it.
