--- src/generators/styles-defaults.ts
@@ -1,1222 +1,0 @@
-import type { Styles } from "./styles-schema";
-
-// The default preset, complete: the single place style defaults exist in the app.
-export const DEFAULT_STYLES: Styles = {
-  map: {
-    attrs: {
-      "background-color": "#000000",
-      filter: null
-    },
-    options: {
-      dataFilter: null
-    }
-  },
-  ocean: {
-    options: {
-      pattern: "./images/pattern1.png",
-      patternOpacity: 0.2
-    },
-    base: {
-      attrs: {
-        fill: "#466eab"
-      }
-    },
-    oceanLayers: {
-      attrs: {
-        filter: null
-      },
-      options: {
-        outline: "-6,-3,-1"
-      }
-    }
-  },
-  landmass: {
-    attrs: {
-      opacity: 1,
-      fill: "#eef6fb",
-      filter: null
-    }
-  },
-  texture: {
-    attrs: {
-      opacity: null,
-      filter: null,
-      mask: "url(#land)"
-    },
-    options: {
-      href: "./images/textures/marble-big.jpg",
-      x: 0,
-      y: 0
-    }
-  },
-  heightmap: {
-    landHeights: {
-      attrs: {
-        opacity: 1,
-        filter: null,
-        mask: "url(#land)"
-      },
-      options: {
-        scheme: "bright",
-        terracing: 0,
-        skip: 5,
-        relax: 0,
-        curve: "curveBasisClosed",
-        render: false
-      }
-    },
-    oceanHeights: {
-      attrs: {
-        opacity: 1,
-        filter: null,
-        mask: null
-      },
-      options: {
-        scheme: "bright",
-        terracing: 0,
-        skip: 0,
-        relax: 1,
-        curve: "curveBasisClosed",
-        render: false
-      }
-    }
-  },
-  biomes: {
-    attrs: {
-      opacity: null,
-      filter: null,
-      mask: "url(#land)"
-    }
-  },
-  cells: {
-    attrs: {
-      opacity: null,
-      stroke: "#808080",
-      "stroke-width": 0.1,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      filter: null,
-      mask: null
-    }
-  },
-  grid: {
-    attrs: {
-      opacity: 0.8,
-      stroke: "#777777",
-      "stroke-width": 0.5,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      transform: null,
-      filter: null,
-      mask: null
-    },
-    options: {
-      type: "pointyHex",
-      scale: 1,
-      dx: 0,
-      dy: 0
-    }
-  },
-  coordinates: {
-    attrs: {
-      opacity: 1,
-      stroke: "#d4d4d4",
-      "stroke-width": 1,
-      "stroke-dasharray": 5,
-      "stroke-linecap": null,
-      filter: null,
-      mask: null
-    },
-    options: {
-      fontSize: 12
-    }
-  },
-  compass: {
-    attrs: {
-      opacity: 0.8,
-      transform: null,
-      filter: null,
-      mask: "url(#water)",
-      "shape-rendering": "optimizespeed"
-    },
-    compassRose: {
-      attrs: {
-        transform: "translate(80 80) scale(0.25)"
-      }
-    }
-  },
-  rivers: {
-    attrs: {
-      opacity: null,
-      fill: "#5d97bb",
-      filter: null
-    }
-  },
-  lakes: {
-    freshwater: {
-      attrs: {
-        opacity: 0.5,
-        fill: "#a6c1fd",
-        "fill-opacity": null,
-        stroke: "#5f799d",
-        "stroke-width": 0.7,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    },
-    salt: {
-      attrs: {
-        opacity: 0.5,
-        fill: "#409b8a",
-        "fill-opacity": null,
-        stroke: "#388985",
-        "stroke-width": 0.7,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    },
-    sinkhole: {
-      attrs: {
-        opacity: 1,
-        fill: "#5bc9fd",
-        "fill-opacity": null,
-        stroke: "#53a3b0",
-        "stroke-width": 0.7,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    },
-    frozen: {
-      attrs: {
-        opacity: 0.95,
-        fill: "#cdd4e7",
-        "fill-opacity": null,
-        stroke: "#cfe0eb",
-        "stroke-width": 0,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    },
-    lava: {
-      attrs: {
-        opacity: 0.7,
-        fill: "#90270d",
-        "fill-opacity": null,
-        stroke: "#f93e0c",
-        "stroke-width": 2,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: "url(#crumpled)"
-      }
-    },
-    dry: {
-      attrs: {
-        opacity: 1,
-        fill: "#c9bfa7",
-        "fill-opacity": null,
-        stroke: "#8e816f",
-        "stroke-width": 0.7,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    }
-  },
-  coastline: {
-    sea_island: {
-      attrs: {
-        opacity: 0.5,
-        stroke: "#1f3846",
-        "stroke-width": 0.5,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      },
-      options: {
-        autoFilter: 1
-      }
-    },
-    lake_island: {
-      attrs: {
-        opacity: 1,
-        stroke: "#7c8eaf",
-        "stroke-width": 0.35,
-        "stroke-dasharray": null,
-        "stroke-linecap": null,
-        filter: null
-      }
-    }
-  },
-  relief: {
-    attrs: {
-      opacity: null,
-      filter: null,
-      mask: null
-    },
-    options: {
-      set: "simple",
-      size: 1,
-      density: 0.4
-    }
-  },
-  religions: {
-    attrs: {
-      opacity: 0.7,
-      stroke: "#777777",
-      "stroke-width": 0,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      filter: null
-    }
-  },
-  cultures: {
-    attrs: {
-      opacity: 0.6,
-      stroke: "#777777",
-      "stroke-width": 0.5,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      filter: null
-    }
-  },
-  states: {
-    statesBody: {
-      attrs: {
-        opacity: 0.4,
-        filter: null
-      }
-    },
-    statesHalo: {
-      attrs: {
-        opacity: 0.4,
-        "stroke-width": 10,
-        filter: "blur(3.5px)"
-      },
-      options: {
-        width: 10
-      }
-    }
-  },
-  provinces: {
-    attrs: {
-      opacity: 0.7,
-      fill: "#000000",
-      "font-size": 10,
-      "font-family": "Georgia",
-      filter: null
-    }
-  },
-  zones: {
-    attrs: {
-      opacity: 0.6,
-      stroke: "#333333",
-      "stroke-width": 0,
-      "stroke-dasharray": null,
-      "stroke-linecap": "butt",
-      filter: null,
-      mask: null
-    }
-  },
-  borders: {
-    stateBorders: {
-      attrs: {
-        opacity: 0.8,
-        stroke: "#56566d",
-        "stroke-width": 1,
-        "stroke-dasharray": 2,
-        "stroke-linecap": "butt",
-        filter: null
-      }
-    },
-    provinceBorders: {
-      attrs: {
-        opacity: 0.8,
-        stroke: "#56566d",
-        "stroke-width": 0.12,
-        "stroke-dasharray": "0.75 0.75",
-        "stroke-linecap": "butt",
-        filter: null
-      }
-    }
-  },
-  routes: {
-    roads: {
-      attrs: {
-        opacity: 0.9,
-        stroke: "#d06324",
-        "stroke-width": 0.7,
-        "stroke-dasharray": 2,
-        "stroke-linecap": "butt",
-        filter: null,
-        mask: null
-      }
-    },
-    trails: {
-      attrs: {
-        opacity: 0.9,
-        stroke: "#d06324",
-        "stroke-width": 0.25,
-        "stroke-dasharray": ".8 1.6",
-        "stroke-linecap": "butt",
-        filter: null,
-        mask: null
-      }
-    },
-    searoutes: {
-      attrs: {
-        opacity: 0.9,
-        stroke: "#ffffff",
-        "stroke-width": 0.35,
-        "stroke-dasharray": "1 2",
-        "stroke-linecap": "round",
-        filter: null,
-        mask: null
-      }
-    }
-  },
-  temperature: {
-    attrs: {
-      opacity: null,
-      fill: "#000000",
-      "fill-opacity": 0.3,
-      stroke: null,
-      "stroke-width": 1.8,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      "font-size": "8px",
-      filter: null
-    }
-  },
-  ice: {
-    attrs: {
-      opacity: 0.9,
-      fill: "#f1f8fe",
-      stroke: "#e8f0f6",
-      "stroke-width": 0.5,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      filter: "url(#dropShadow01)"
-    }
-  },
-  precipitation: {
-    attrs: {
-      opacity: null,
-      fill: "#003dff",
-      stroke: "#000000",
-      "stroke-width": 0,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      filter: null
-    }
-  },
-  population: {
-    attrs: {
-      opacity: null,
-      "stroke-width": 1.6,
-      "stroke-dasharray": null,
-      "stroke-linecap": "butt",
-      filter: null
-    },
-    rural: {
-      attrs: {
-        stroke: "#0000ff"
-      }
-    },
-    urban: {
-      attrs: {
-        stroke: "#ff0000"
-      }
-    }
-  },
-  emblems: {
-    attrs: {
-      opacity: 0.9,
-      "stroke-width": 1,
-      filter: null
-    },
-    stateEmblems: {
-      options: {
-        size: 1
-      }
-    },
-    provinceEmblems: {
-      options: {
-        size: 1
-      }
-    },
-    burgEmblems: {
-      options: {
-        size: 1
-      }
-    }
-  },
-  labels: {
-    attrs: {
-      "font-size": "100px"
-    },
-    groups: {
-      capital: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "6%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.5
-        }
-      },
-      city: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "5%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.4
-        }
-      },
-      fort: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "2%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.5
-        }
-      },
-      monastery: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "2%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.5
-        }
-      },
-      caravanserai: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "2%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.5
-        }
-      },
-      trading_post: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "2%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.5
-        }
-      },
-      village: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "3%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.4
-        }
-      },
-      hamlet: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "2%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.4
-        }
-      },
-      town: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": null,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "4%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: -0.4
-        }
-      },
-      state: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: "#3a3a3a",
-          "stroke-width": 0,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "22%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: 0
-        }
-      },
-      province: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: "#3a3a3a",
-          "stroke-width": 0,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "6%",
-          "font-family": "Cinzel",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: 0
-        }
-      },
-      route: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": 0,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "1.5%",
-          "font-family": "Lugrasimo",
-          style: "text-shadow: white 0px 1px 6px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: 0
-        }
-      },
-      river: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: null,
-          "stroke-width": 0,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "1.5%",
-          "font-family": "Lugrasimo",
-          style: "text-shadow: white 0px 1px 6px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: 0
-        }
-      },
-      added: {
-        attrs: {
-          opacity: 1,
-          fill: "#3e3e4b",
-          "fill-opacity": null,
-          stroke: "#3a3a3a",
-          "stroke-width": 0,
-          "stroke-dasharray": null,
-          "stroke-linecap": null,
-          "letter-spacing": 0,
-          "font-size": "18%",
-          "font-family": "Almendra SC",
-          style: "text-shadow: white 0px 0px 4px",
-          filter: null
-        },
-        options: {
-          dx: 0,
-          dy: 0
-        }
-      }
-    }
-  },
-  burgIcons: {
-    burgIcons: {
-      groups: {
-        capital: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 2,
-            icon: "#icon-square"
-          }
-        },
-        city: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 1.5,
-            icon: "#icon-circle"
-          }
-        },
-        fort: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-square"
-          }
-        },
-        monastery: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-cross"
-          }
-        },
-        caravanserai: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-triangle"
-          }
-        },
-        trading_post: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-triangle"
-          }
-        },
-        village: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        hamlet: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 0.5,
-            icon: "#icon-circle"
-          }
-        },
-        town: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": 0.7,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": "butt",
-            "stroke-linejoin": "round",
-            filter: null
-          },
-          options: {
-            size: 1,
-            icon: "#icon-circle"
-          }
-        }
-      }
-    },
-    anchors: {
-      groups: {
-        capital: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 1.9,
-            icon: "#icon-circle"
-          }
-        },
-        city: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 1.5,
-            icon: "#icon-circle"
-          }
-        },
-        fort: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        monastery: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        caravanserai: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        trading_post: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        village: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.7,
-            icon: "#icon-circle"
-          }
-        },
-        hamlet: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 0.5,
-            icon: "#icon-circle"
-          }
-        },
-        town: {
-          attrs: {
-            opacity: 1,
-            fill: "#ffffff",
-            "fill-opacity": null,
-            stroke: "#3e3e4b",
-            "stroke-width": 1.2,
-            "stroke-dasharray": null,
-            "stroke-linecap": null,
-            "stroke-linejoin": null,
-            filter: null
-          },
-          options: {
-            size: 1,
-            icon: "#icon-circle"
-          }
-        }
-      }
-    }
-  },
-  goods: {
-    goodsCells: {
-      attrs: {
-        opacity: 1,
-        filter: null
-      }
-    },
-    goodsIcons: {
-      attrs: {
-        opacity: 1,
-        "stroke-width": 0.3,
-        filter: "url(#dropShadow01)"
-      },
-      options: {
-        size: 6,
-        circle: true
-      }
-    },
-    goodsBurgs: {
-      attrs: {
-        opacity: 1,
-        stroke: "#41414f",
-        "stroke-width": 0.2,
-        filter: null
-      },
-      options: {
-        size: 3
-      }
-    }
-  },
-  markets: {
-    attrs: {
-      opacity: 1,
-      fill: null,
-      "fill-opacity": 0.03,
-      "stroke-width": 1,
-      "stroke-opacity": 0.8,
-      filter: null
-    },
-    options: {
-      size: 3,
-      fontSize: 5,
-      icon: "⚖️"
-    }
-  },
-  trade: {
-    attrs: {
-      opacity: null,
-      filter: null
-    }
-  },
-  markers: {
-    attrs: {
-      opacity: null,
-      filter: null
-    },
-    options: {
-      rescale: 1
-    }
-  },
-  military: {
-    attrs: {
-      opacity: null,
-      stroke: "#000",
-      "stroke-width": 0.3,
-      "stroke-dasharray": null,
-      "stroke-linecap": null,
-      "fill-opacity": 1,
-      filter: null
-    },
-    options: {
-      fontSize: 6,
-      boxSize: 3
-    }
-  },
-  rulers: {
-    attrs: {
-      opacity: null,
-      "stroke-width": 2,
-      "stroke-dasharray": "10",
-      "stroke-linecap": null,
-      filter: null
-    },
-    options: {
-      fontSize: 20
-    }
-  },
-  scaleBar: {
-    attrs: {
-      opacity: 1,
-      fill: "#353540",
-      "font-size": 10
-    },
-    options: {
-      barSize: 2,
-      x: 99,
-      y: 99,
-      label: ""
-    },
-    back: {
-      attrs: {
-        opacity: 0.2,
-        fill: "#ffffff",
-        "fill-opacity": null,
-        stroke: "#000000",
-        "stroke-width": 1,
-        filter: "url(#blur5)"
-      },
-      options: {
-        top: 20,
-        right: 15,
-        bottom: 15,
-        left: 10
-      }
-    }
-  },
-  legend: {
-    attrs: {
-      stroke: "#812929",
-      "stroke-width": 2.5,
-      "stroke-dasharray": "0 4 10 4",
-      "stroke-linecap": "round",
-      "font-family": "Almendra SC"
-    },
-    options: {
-      fontSize: 13,
-      x: 99,
-      y: 93,
-      columns: 8
-    },
-    box: {
-      attrs: {
-        fill: "#ffffff",
-        "fill-opacity": 0.8
-      }
-    }
-  },
-  fogging: {
-    attrs: {
-      opacity: 0.98,
-      fill: "#30426f",
-      mask: "url(#fog)",
-      filter: null
-    }
-  },
-  vignette: {
-    attrs: {
-      opacity: 0.3,
-      fill: "#000000",
-      mask: "url(#vignette-mask)",
-      filter: null
-    },
-    options: {
-      x: "0.3%",
-      y: "0.4%",
-      width: "99.6%",
-      height: "99.2%",
-      rx: "5%",
-      ry: "5%",
-      filter: "blur(20px)"
-    }
-  }
-};
